mlxsw: spectrum: Support IPIP underlay VRF migration
When a bound device of a tunnel netdevice changes VRF, the offload needs
to be treated much the same way as when the tunnel netdevice itself
changes VRF.

The difference is that encapsulating routes remain valid. The impacted
next hops need to be moved over to the new RIF, and the impacted next
hop groups refreshed.

Signed-off-by: Petr Machata <petrm@mellanox.com>