selftests: mlxsw: Add a test for UC behavior under MC flood
A so-called "MC-aware" mode has recently been enabled in mlxsw. In
MC-aware mode, BUM traffic is handled in a special way so that when a
switch is flooded with BUM, UC performance isn't unduly impacted.
Without enablement of this mode, a stream of BUM traffic can cause
sustained UC throughput drop in excess of 99 %.

Add a test for this handling. Compare how much UC throughput degrades as
a stream of broadcast frames floods the switch. Some degradation is
expected (there's more traffic to handle, after all), but make sure it's
not pathological.

Signed-off-by: Petr Machata <petrm@mellanox.com>