A sketch of RED selftest
This sends traffic at 1Gbps to clog an egress port ("heavy stream") in
through port 1, and then injects a light stream of traffic through port 2
to create a backlog at port 3. Both light and heavy are ECN enabled TCP
traffic. On egress port 3, RED is configured to mark ECN traffic as ECN-CE
if the backlog is too large.

The overall diagram is approximately:
           __________
	  |          |
 heavy >---1         |
          |         3---> result
 light >---2         |
          |__________|

In order to check integration of RED with PRIO / ECN, 1, 2 and 3 are
actually dual VLAN ports. Traffic on VLAN 10 is assigned priority 0,
traffic on VLAN 11 is assigned priority 1. PRIO or ECN at port 3 has
priomap to direct prio 0 to band 7 (thus TC 0), and prio 1 to band 6 (thus
TC 1).

The heavy stream is:

$ mausezahn swp5.10 -p 8000 -c 0 -A 192.0.2.1 -B 192.0.2.4 \
    -b 7c:fe:90:f5:a3:76 -t tcp tos=0x01

And the light stream:

$ mausezahn swp8.10 -p 8000 -d 10m -c 0 -A 192.0.2.3 -B 192.0.2.4 \
    -b 7c:fe:90:f5:a3:76 -t tcp tos=0x01

Result traffic can then be analyzed by installing a flower rule to note
ECN-CE traffic:

$ tc f a dev swp10 ingress pref 101 protocol 802.1q flower \
    vlan_ethtype ipv4 ip_tos 0x03/0x03 action pass

The queue size can be seen in tc_transmit_queue_tc_0 and _1. As the queue
builds up, for traffic in tc_0 we would expect to see ECN-CE go to 1Gbps
around the 300KB mark. For tc_1, around 500KB. (As per current settings,
this can be changed at RED Qdisc.)

The delay parameter of the light stream needs to be tweaked to get a
reasonable buildup speed, so that the point at which ECN marking occurs can
be determined with any precision. Doing this in automated fashion is
probably going to be error-prone.

I find that when one sends two streams of heavy traffic to one 1Gbps
port (and configures shared buffer so as not to take all the memory in the
system), that then creates a much more robust heavy stream (which is still
1Gbps):
            __________               __________
	   |          |             |	       |
 heavy1 >---1        3---> heavy >---4        6---> result
           |          |             |          |
 heavy2 >---2         |    light >---5         |
           |__________|             |__________|

With this setup, every packet of light traffic goes directly to the
backlog. This allows a much more accurate buffer buildup: if I want to
check that no ECN-CE marking encounters at 250KB, I just generate 250KB of
traffic, and check the flower counters. Then I send in another 100KB and
check again, expecting now to get 1Gbps of ECN-CE traffic.

This however needs 10 ports (heavy1, heavy2, light, result hosts and 6
numbered switch ports), and our systems commonly only have 8 looped-back
ports. Some something clever with reusing result host to send the light
stream needs to be invented to selftestize this. I was using VMs to provide
the heavy traffic.