af_netlink: Fix shift out of bounds in group mask calculation
When a netlink message is received, netlink_recvmsg() fills in the address
of the sender. One of the fields is the 32-bit nl_groups, which carries the
multicast group, on which the message was received. Since this field is
only 32 bits wide, it cannot actually carry group numbers above 31. And the
UB sanitizer flags the out-of-bounds shift attempts.

Which bits end up being set is implementation defined, but it's either
going to be a wrong non-zero value, or zero, which is arguably wrong as
well, but at least with a bitfield it is obviously wrong. Make the latter
choice deterministic by always setting to 0 for higher-numbered multicast
groups.

The way to trigger this issue is e.g. through monitoring the BRVLAN group:

	# bridge monitor vlan &
	# ip link add name br type bridge

Which produces the following citation:

	UBSAN: shift-out-of-bounds in net/netlink/af_netlink.c:162:19
	shift exponent 32 is too large for 32-bit type 'int'

Signed-off-by: Petr Machata <petrm@nvidia.com>