mlxsw: Configure descriptor buffers
Spectrum machines have two resources related to keeping packets in an
internal buffer: bytes (allocated in cell-sized units) for packet payload,
and descriptors, for keeping headers. Currently, mlxsw only configures the
bytes part of the resource management.

However, certain workloads consist of small packets, and in that case,
configuring the shared buffer to be able to take the traffic itself is not
enough, because no-buffer drops will occur due to exhaustion of the
descriptor buffer.

Spectrum switches permit a full parallel configuration for the descriptor
resources, including port-pool and port-TC-pool quotas. By default, these
are all configured to use pool 14, with an infinite quota. The ingress pool
14 is then infinite in size.

However, egress pool 14 has a finite size, which is chip-dependent, but in
all cases much lower than what the chip actually permits. As a result, we
can easily construct workloads that exhaust the configured descriptor
limit.

There is no good reason not to use the resources that are available to the
chip. Therefore configure the egress descriptor pool to be infinite in size
as well.

In the code, include both the configuration of ingress and ingress, mostly
for clarity.

Signed-off-by: Petr Machata <petrm@nvidia.com>