mlxsw: spectrum_trap: Register 802.1X packet traps with devlink
Register the previously added packet traps with devlink. This allows
user space to tune their policers and in the case of the FDB mismatch
trap, user space can set its action to "trap" in order to gain
visibility into packets that were discarded by the device due to an FDB
mismatch.

The two trap groups are initialized with identical properties to similar
groups, such as STP and layer 3 exceptions.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>
Reviewed-by: Petr Machata <petrm@nvidia.com>