mlxsw: Extend PGT APIs to support maintaining list of ports per entry
Add the following APIs to associate a PGT entry with SMPE index and add
or remove a port:
1. mlxsw_sp_pgt_entry_port_set() will be used by FID code, to add/remove
   port from specific PGT entry.
2. mlxsw_sp_pgt_entry_port_bitmap_set() will be used by MDB code, to
   add/remove several local ports from a specific PGT entry. A bitmap of
   ports will be used.

When the first port is added to PGT entry, allocate the entry in the given
MID index, when the last port is removed from PGT entry, free it.

The ports are reference counted per PGT entry as multiple MDB entries can
use the same hardware MDB entry. It occurs because MDB entries in the Linux
bridge are keyed according to their multicast IP, when these entries are
notified to device drivers via switchdev, the multicast IP is converted to
a multicast MAC. This conversion might cause collisions, for example,
ff0e::1 and ff0e:1234::1 are both mapped to the multicast MAC
33:33:00:00:00:01.

Update the PGT table in hardware only when port is first added or removed,
according to the reference counting.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>