mlxsw: spectrum_router: Maintain a hash table of CRIFs
CRIFs are objects that mlxsw maintains for netdevices that may not have an
associated RIF (i.e. they may not have been instantiated in the ASIC), but
if indeed they do not, it is quite possible they will in the future. These
netdevices are candidate RIFs, hence CRIFs. Netdevices for which CRIFs are
created include e.g. bridges, LAGs, or front panel ports. The idea is that
next hops and neighbors would be kept at CRIFs, not RIFs, and thus make it
possible to easily offload and unoffload the entities that have been added
before the RIF was created.

In this patch, add the code for low-level CRIF maintenance: create and
destroy, and keep in a table keyed by the netdevice pointer for easy
recall.

Further, install a new netdevice notifier that processes the REGISTER and
UNREGISTER events and creates and destroys CRIFs as appropriate. The reason
a new notifier is called for is that the lifetime of CRIFs is larger than
the lifetime of RIFs. Therefore the netdevice notifier needs to be
installed fairly soon. However processing many events requires that the
router be fully initialized, which it cannot be yet. Hence, a new notifier
that only handles the CRIF lifetime.

Signed-off-by: Petr Machata <petrm@nvidia.com>