mlxsw: mlxsw_sp_router_schedule_work(): Move here mlxsw_sp_netevent_work.n init
After the struct mlxsw_sp_netevent_work.n field initialization is moved
here, the body of code that handles NETEVENT_NEIGH_UPDATE is almost
identical to the one in the helper function. Therefore defer to the helper
instead of inlining the equivalent.

The one difference is that previously, the mlxsw_sp_port_lower_dev_hold()
made sure that the work was only scheduled for physical netdevice uppers.
However, the neighbor management code is currently dependent on existence
of a RIF, not of a front panel lower, and handles the case of a missing RIF
correctly.

Signed-off-by: Petr Machata <petrm@nvidia.com>