mlxsw: spectrum_router: Track next hops at CRIFs
Move the list of next hops from mlxsw_sp_rif to mlxsw_sp_crif. The reason
is that eventually, next hops for mlxsw uppers should be offloaded and
unoffloaded lazily as a netdevice becomes an upper, or stops being one.
Currently, next hops are tracked at RIFs, but RIFs do not exist when a
netdevice is not an mlxsw uppers. CRIFs are kept track of throughout their
lifetime.

Correspondingly, track at each next hop not its RIF, but its CRIF (from
which a RIF can be deduced).

Note that now that next hops are tracked at CRIF, it is not necessary to
move each over to the new RIF. mlxsw_sp_nexthop_rif_migrate() is therefore
gone and mlxsw_sp_rif_migrate_destroy() calls mlxsw_sp_nexthop_rif_update()
directly.

Signed-off-by: Petr Machata <petrm@nvidia.com>