mlxsw: spectrum_router: Split nexthop finalization to RIF/CRIF stages [xxx]
Nexthop finalization consists of two steps: the part where the offload is
removed, because the backing RIF is now gone; and the part where the
association to the CRIF is severed, because the CRIF is getting removed.

[xxx] CRIF is removed on netdev unregistration. Shouldn't the next hops be
gone by then?

Extract from mlxsw_sp_nexthop_type_fini() a helper that covers the RIF-gone
part, mlxsw_sp_nexthop_type_rif_gone(), so that it can later be called
independently.

Signed-off-by: Petr Machata <petrm@nvidia.com>