mlxsw: spectrum_switchdev: Create a RIF when PVID is added
Currently, addition of an IP address to a bridge with a PVID prompts
creation of a RIF for that bridge / VLAN. Addition of a PVID to a bridge
that already has an address should similarly cause RIF to be created. Add
the necessary logic.

Now that it is not necessary to enforce ordering of IP address and PVID
addition, stop forbidding addition of an IP address to a PVID-less bridge.
Instead, bail out quietly.

Signed-off-by: Petr Machata <petrm@nvidia.com>