mlxsw: spectrum: Permit enslavement to bridges with uppers [xxx]
Enslaving of front panel ports (and their uppers) to netdevices that
already have uppers is currently forbidden. mlxsw currently validates each
upper change, but only considers those that pertain to "interesting"
netdevices (notably bridges that have front panel ports as lowers). Thus
when the attachment to netdevices with uppers should be allowed, all these
previously ignored notifiers need to be reconsidered.

To that end, invoke the CHANGEUPPER event handler for each netdevice above
the one that the front panel port is being enslaved to.

Signed-off-by: Petr Machata <petrm@nvidia.com>