mlxsw: spectrum_router: Extract a helper to free a RIF
Right now freeing the object that mlxsw uses to keep track of a RIF is as
simple as calling a kfree. But later on as CRIF abstraction is brought in,
it will involve severing the link between CRIF and its RIF as well. Better
to have the logic encapsulated in a helper.

Signed-off-by: Petr Machata <petrm@nvidia.com>