mlxsw: spectrum_router: Use router.lb_crif instead of .lb_rif_index [xxx]
A previous commit added a pointer to loopback CRIF to the router data
structure. That makes the loopback RIF index redundant, as everything
necessary can be derived from the CRIF. Drop the field and adjust the code
accordingly.

 x x x x x x xxx this is the old commit message:

Despite dropping traffic instead of forwarding it, black-hole next hops
nonetheless need to point to a valid RIF. Currently this is handled by
getting an underlay RIF reference during router init. This then primes
struct mlxsw_sp_router.lb_rif_index, which is used to look up the RIF when
initializing black-hole next hop's RIF pointer.

Eventually, we would like to track next hops at CRIFs, not RIFs though, and
that means there needs to be a CRIF where the next hop would be tracked,
and that the next hop would point at.

Core kernel black-hole next hops do reference a netdevice though, namely
the loopback netdevice for the given netns (see nexthop_create()). We can
thus do the same: keep a CRIF for the netns's loopback.

In this patch, replace the lb_rif_index with a CRIF pointer, constructed
during router init. Extend several helpers to be able to connect the
created loopback RIF with this CRIF.

[xxx] in code

Signed-off-by: Petr Machata <petrm@nvidia.com>