net: bridge: Maintain number of MDB entries in net_bridge_mcast_port
At this point, there is a bridge-global limit on the number of entries that
MDB is allowed to hold, struct net_bridge::hash_max.

In order to provide a more fine-grained limitation knobs, this patch
extends struct net_bridge_mcast_port with two fields: the current number of
MDB entries, and a maximum number permitted (or 0 for no limit).

The per-port multicast context is used for tracking of MDB entries for the
port as a whole. This is used for bridges that have VLAN filtering off, or
that have it on, but have mcast_vlan_snooping off.

The per-port-VLAN multicast context is then used on VLAN-filtering bridges
on VLANs that have multicast snooping on.

With these changes in place, it will be possible to configure MDB limit for
bridge as a whole, or any one port as a whole, or any single port-VLAN.

In this patch, introduce only the accounting for number of entries, and the
max field itself, but not the means to toggle the max. The next patch
introduces the netlink APIs to toggle and read the values.

Signed-off-by: Petr Machata <petrm@nvidia.com>