net: dcb: Clean up clean-up paths in dcbnl_cee_fill()
This function opens a nest near the beginning by calling
nla_nest_start_noflag(). No error path cancels this nest though.
Instead, the cleanup block at the end of the function just sets
err and returns it immediately.

This is not a bug per se: the caller, having seen an error, trims the
message at the header by calling nlmsg_cancel(), thus trimming the nest as
well. However, these dangling nests make it hard to reason about the
code.

Therefore, in error handlers where the nest is not open, return -EMSGSIZE
right away. Reuse the cleanup block block to cancel the nest, and return a
fixed -EMSGSIZE instead of passing it through the err variable. Drop err
initialization, which was never necessary.

Signed-off-by: Petr Machata <petrm@nvidia.com>