net: dcb: Clean up clean-up paths in dcbnl_getnumtcs()
This function opens a nest for DCB_ATTR_NUMTCS by calling
nla_nest_start_noflag(). The error path does not cancel this nest though.

This is not a bug per se. This function is called through a pointer from
dcb_doit(), and on error, nlmsg_free() gets rid of the whole message,
including the nest. However, these dangling nests make it hard to reason
about the code (especially since the call chain is not obvious due to the
indirect calls).

Thus invoke nla_nest_cancel() before returning the error.

Signed-off-by: Petr Machata <petrm@nvidia.com>