net: dcb: Convert dcbnl_build_peer_app() to a more idiomatic style
Instead of the more usual style that goes like this:

	err = foo();
	if (err)
		goto out;

This function is written using this style:

	err = foo();
	if (!err)
		do_stuff;

This makes it a bit trickier to clean up error paths in. Convert it to the
former style to make it easier to add an nla_nest_cancel() call that the
function is missing.

Signed-off-by: Petr Machata <petrm@nvidia.com>