net: ipv4: Allow changing IPv4 labels
IPv4 addresses can be tagged with label strings. Unlike IPv6 addrlabels,
which are used for prioritization of IPv6 addresses, these "ip address
labels" are simply tags that the userspace can assign to IP addresses
arbitrarily.

IPv4 has had support for these tags since before Linux was tracked in GIT.
However it has never been possible to change the label after it is once
defined. This limits usefulness of this feature. The userspace technically
still can just put arbitrary label to a given address, but needs to delete
and recreate it with the proper label, which is impractical and disrupts
routing.

In this patchset, when an address is replaced and a label is given, update
the label at the address to the one given in the request. So far the new
value was simply ignored. Technically this constitutes an API breakage, but
it is hard to imagine that a RTM_NEWADDR request with NLM_F_REPLACE flag
and carrying an IFA_LABEL attribute should mean anything else that yes,
that label should be replaced.

Signed-off-by: Petr Machata <petrm@nvidia.com>