selftests: forwarding: hw_stats_l3: Detect failure to install counters
Running this test makes little sense if the enabled l3_stats are not
actually reported as "used". This can signify a failure of a driver to
install the necessary counters, or simply lack of support for enabling
counters on a given netdevice. It is generally impossible to tell from the
outside which it is. But more likely than not, if somebody is running this
on veth pairs, they likely do not intend to actually test that a certain
piece of HW can install in-HW counters corresponding to these netdevices.
It is more likely they are running the test by mistake or similar.

Therefore detect that the counter has not been actually installed. If the
netdevice is one end of a veth pair, SKIP. Otherwise FAIL.

Reported-by: Danielle Ratson <danieller@nvidia.com>
Suggested-by: Hangbin Liu <liuhangbin@gmail.com>
Signed-off-by: Petr Machata <petrm@nvidia.com>