net: switchdev: Add a helper to replay objects on a bridge port
When a front panel joins a bridge via another netdevice (typically a LAG),
the driver needs to learn about the objects configured on the bridge port.
When the bridge port is offloaded by the driver for the first time, this
can be achieved by passing a notifier to switchdev_bridge_port_offload().
The notifier is then invoked for the individual objects (such as VLANs)
configured on the bridge, and can filter out the interesting ones.

Doing this when the second port joins the bridge lower is unnecessary,
but the replay is still needed. To that end, add a new function,
switchdev_bridge_port_replay(), which does only the replay part of
the _offload() function in exactly the same way as that function.

Signed-off-by: Petr Machata <petrm@nvidia.com>
Reviewed-by: Danielle Ratson <danieller@nvidia.com>