mlxsw: spectrum_nve: Use tracker helpers to hold & put netdevices
Using the tracking helpers makes it easier to debug netdevice refcount
imbalances when CONFIG_NET_DEV_REFCNT_TRACKER is enabled.

In mlxsw_sp_nve_fid_disable(), the reference comes from dev_get_by_index().
Therefore allocate the tracker explicitly. Both get and put are in the same
function, so the tracker is put on stack. This is unlikely to be
practically useful, because the references are just several lines apart,
but makes sense for consistency.

Signed-off-by: Petr Machata <petrm@nvidia.com>