mlxsw: core, pci: Add plumbing related to CFF mode [xxx]
CFF mode, for Compressed FID Flooding, is a way to organize where flood
vectors for individual FIDs should be found in the PGT table. The bus
module determines whether CFF is supported, can configure flood mode to CFF
if it is, and knows what flood mode has been configured. Therefore add a
bus callback to determine the configured flood mode. Also add to core an
API to query it.

The CFF mode is for not used. The code to actually use it will be added
later.

Signed-off-by: Petr Machata <petrm@nvidia.com>