mlxsw: reg: Set NVE flood profile as part of FID configuration
The NVE flood profile is used for determining of offset applied to KVD for
NVE flood. We currently do not set it, leaving it at the default value of
0. In the CFF mode, we will need to use more profiles, and it will make
sense to keep the profile zero as unused, to spot configuration errors.

Therefore add the corresponding argument to mlxsw_reg_sfmr_pack(). Add a
new enum for flood profiles, with one member for the NVE profile. Pass that
profile to the adapted pack function.

Offsets for individual profile_id's and traffic types are configured by
SFFP. However the default offset is 0, which is fine for the NVE use,
and we can therefore leave it unconfigured for now.

Signed-off-by: Petr Machata <petrm@nvidia.com>