mlxsw: reg: Add new fields to the SFMR register
Add the field cff_mid_base, which specifies at which point in PGT the
per-FID flood table is stored. Ad cff_prf_id, the profile ID, which
determines on which row of the flood table a flood vector can be found for
a given traffic type.

Also add a pack function for the CFF mode, which makes use of the new
fields.

Signed-off-by: Petr Machata <petrm@nvidia.com>