mlxsw: spectrum_fid: Privatize the Spectrum-2 FID family array
Currently, mlxsw always uses a "controlled" flood mode on all Nvidia
Spectrum generations. The following patches will however introduce a
possibility to run a "CFF" (for Compressed FID Flooding) mode on newer
machines, if the FW supports it.

Several operations will differ between how they need to be done in
controlled mode vs. CFF mode. Thus the per-FID-family ops will differ
between controlled and CFF, thus the FID family array as such will
differ depending on whether the mode negotiated with FW is controlled
or CFF.

It is thus necessary to introduce a bit of logic that decides which array
to use. Thus hide the Spectrum>=2 array and introduce a function to yield
it instead.

Signed-off-by: Petr Machata <petrm@nvidia.com>