mlxsw: spectrum_fid: Split a helper out of mlxsw_sp_fid_flood_table_mid()
In future patches, for CFF flood mode support, we will need a way to
determine a PGT base dynamically, as an op. Therefore, for symmetry,
split out a helper, mlxsw_sp_fid_ctl_pgt_base(), that determines a PGT base
in the controlled mode as well.

Note that the operation to determine a PGT base is fundamentally flood mode
specific: in controlled mode, the PGT base is determined from the FID
family and flood table, and FID indexes into that table. In CFF mode, it's
determined from the FID family and FID index, and the flood table indexes
into that table. Thus they are two distinct operations: ctl_pgt_base and
cff_pgt_base. Reflect that in the name of the helper.

Now that the helper is available, use it in mlxsw_sp_fid_flood_table_init()
which currently invokes the FID->MID helper to that end.

Signed-off-by: Petr Machata <petrm@nvidia.com>