mlxsw: spectrum_fid: Add an op for packing SFMR
The way SFMR is packed differs between the controlled and CFF flood modes.
Add an op to dispatch it dynamically.

In CFF mode, for rFIDs, the operation will involve querying spectrum core.
This operation ought not fail in practice, but in principle is fallible and
returns an error code. I find it cleaner and more future-proof to propagate
the error code through the pack operation upwards. Thus make the op return
an int, and recognize the possible error in the sole caller.

Signed-off-by: Petr Machata <petrm@nvidia.com>